\name{ackroyd}
\alias{ackroyd}
\title{Sales and related data from the fictional town of Ackroyd.}
\description{
  Data from six fictional firms in Ackroyd. Rows are firms; columns of data are named as follows:
    \itemize{
      \item{"firm"}{Firm number (1 to 6)}
      \item{"mplyees"}{Number of employees in the firm}
      \item{"sales"}{Annual sales in hundreds of thousands of pounds}
      \item{"hire"}{Intention to hire more staff: 1=Yes, 0=No}
      \item{"location"}{1=east side of town, 2=west side of town}
      \item{"nature"}{1=manufacturing, 2=non-manufacturing}
    }
}
\usage{data(ackroyd)}
\format{A data frame}
\source{Tryfos (1996)}
\references{Tryfos, P. 1996. Sampling methods for applied research. New York: John Wiley and Sons.}
\keyword{datasets}
