\name{airport}
\alias{airport}
\title{US airport data.}
\description{
  Statistics on departures and arrivals at 135 US airports. Rows are %%@
    airports; columns of data are named as follows:
    \itemize{
      \item{"airport"}{Airport number: 1 to 135}
      \item{"schdep"}{Number of scheduled depatrures}
      \item{"actdep"}{Number of actual depatrures}
      \item{"passngrs"}{Number of passengers passing through airport}
      \item{"freitrev"}{Freight revenue in thousands of dollars}
      \item{"mailrev"}{Mail revenue in thousands of dollars}
      \item{"name"}{Name of airport}
      \item{"city"}{City containing airport}
    }
}
\usage{data(airport)}
\format{A data frame}
\source{forgotten}
\references{None}
\keyword{datasets}