% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algofastflightcube.R
\name{algofastflightcubeSPOT}
\alias{algofastflightcubeSPOT}
\title{algofastflightcubeSPOT}
\usage{
algofastflightcubeSPOT(X, pik)
}
\arguments{
\item{X}{matrix of auxiliary variables.}

\item{pik}{vector of inclusion probabilities}
}
\value{
updated pik
}
\description{
implementation modified from the package sampling.
}
