% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{all_sug}
\alias{all_sug}
\title{all test condition in Rcpp}
\usage{
all_sug(x)
}
\arguments{
\item{x}{vector of boolean}
}
\value{
true or false
}
\description{
all test condition in Rcpp
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
