\name{barnett}
\alias{barnett}
\title{Dataset gathered by Barnett from his sampling theory students.}
\description{
  Some seemingly arbitray data gathered from a stats class to illustrate %%@
    sampling theory methods. Rows are students; columns of data are named %%@
    as follows:
    \itemize{
      \item{"student"}{Student number}
      \item{"height"}{measured in inches, above a reference height}
      \item{"weight"}{measured in pounds, above a reference weight}
      \item{"sex"}{Sex: 1=Male, 2=Female}
      \item{"row"}{Row student sat in: 1 to 5}
      \item{"column"}{Column student sat in: 1 to 5}
    }
}
\usage{data(barnett)}
\format{A data frame}
\source{Barnett (1991)}
\references{Barnett, V. 1991. Sample survey principles and methods. %%@
              Edward Arnold: London.}
\keyword{datasets}