\name{billionaire}
\alias{billionaire}
\title{Data on billionaires}
\description{
  Statistics on 225 billionaires. Rows are billionares; columns of data 
    are named as follows:
    \itemize{
      \item{"billionaire"}{billionaire number (1 to 225)}
      \item{"wealth"}{How many billion dollars the person is worth}
      \item{"age"}{Billionaire age}
                   \item{``region''}{Region of the world where billionaire is resident}
      }
    }
                   \usage{data(billionaire)}
                   \format{A data frame}
                   \source{forgotten}
                   \references{None}
                   \keyword{datasets}
                   
                   
                   
                   