\name{brexitsample}
\alias{brexitsample}
\title{Sample of voters from the Brexit referendum}
\description{
  A data frame with each row corresponding to a region of the UK. 
  Columns are as follows:
    \itemize{
      \item{"Region"}{ Geographic region}
      \item{"Area"}{ Geographic area}
      \item{"Electorate"}{ Number of people eligible to vote}
      \item{"Valid_Votes"}{ Number of valid votes cast}
      \item{"Remain"}{ Number of votes to remain in the EU}
      \item{"Leave"}{ Number of votes to leave the EU}
      \item{"median_hourly_pay2005"}{ Median hourly pan in 2005}
      \item{"mean_hourly_pay_growth"}{ Mean growth in hourly pay}
      }
    }
\usage{data(brexitsample)}
\format{A data frame}
\source{online}
\keyword{datasets}