% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{check_equal}
\alias{check_equal}
\title{Check equal}
\usage{
check_equal(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{vector}
}
\value{
return true if vector equal to each other
}
\description{
Check equal
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
