% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{choose}
\alias{choose}
\title{Binomial coefficient}
\usage{
choose(n, k)
}
\arguments{
\item{n}{integer.}

\item{k}{integer.}
}
\value{
binomial coefficient
}
\description{
Binomial coefficient
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
