\name{classeg}
                   \alias{classeg}
                   \title{Example dataset used for illustration in Sampling Theory %%@
                   notes.}
                   \description{
                   Test marks and related data from six students. Rows are students; %%@
                   columns of data are named as follows:
                   \itemize{
                   \item{"student"}{Student number}
                   \item{"mark"}{Test mark}
                   \item{"oldgrade"}{Mark on previous test (out of 14)}
                   \item{"sex"}{Sex: 1=Male, 0=Female}
                   \item{"row"}{Row student sat in during test: 1 to 4}
                   }
                   }
                   \usage{data(classeg)}
                   \format{A data frame}
                   \source{None}
                   \references{None}
                   \keyword{datasets}