% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{colSumsRcpp}
\alias{colSumsRcpp}
\title{colSums in Rcpp}
\usage{
colSumsRcpp(x)
}
\arguments{
\item{x}{Matrix X}
}
\value{
colSums
}
\description{
colSums in Rcpp
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
