% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{duplicatedCols}
\alias{duplicatedCols}
\title{duplicated columns}
\usage{
duplicatedCols(B)
}
\arguments{
\item{B}{Matrix B}
}
\value{
return vector of boolean
}
\description{
duplicated columns
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
