% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{duplicatedRows}
\alias{duplicatedRows}
\title{duplicated rows}
\usage{
duplicatedRows(B)
}
\arguments{
\item{B}{Matrix B}
}
\value{
vector of boolean
}
\description{
duplicated rows
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
