% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_flight_cube.R
\name{fast.flight.cube}
\alias{fast.flight.cube}
\title{Title}
\usage{
fast.flight.cube(X, pik, deepness = 1, EPS = 1e-07)
}
\arguments{
\item{EPS}{}
}
\value{

}
\description{
Title
}
\examples{

rm(list = ls())
library(sampling)
library(MASS)
EPS=0.0000001
N=1000
n=300
p=1
q=7
z=runif(N)
#z=rep(1,N)
pik=inclusionprobabilities(z,n)
X=cbind(pik,matrix(rnorm(N*p),c(N,p)))
A=X/pik

Z=cbind(matrix(rbinom(N*q,1,1/2),c(N,q)))
B=cbind(Z,-Z)
r=c(ceiling(pik\%*\%B))
r[abs(pik\%*\%B-round(pik\%*\%B))<EPS]=round(pik\%*\%B)[abs(pik\%*\%B-round(pik\%*\%B))<EPS]
s=fast.flight.cube.ineq(X,pik,B,r,deepness=1,EPS=0.000001)
round(s,3)
c(t(B)\%*\%pik)
c(r)
c(t(B)\%*\%s)
colSums(X)
c(t(X)\%*\%(s/pik))

piks=fast.flight.cube(cbind(X,B*pik),pik,deepness=1,EPS=0.000001)
pikstar=fast.flight.cube.ineq(X/pik*piks,piks,B,r,deepness=1,EPS=0.000001)
}
