% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastflightcubeSPOT.R
\name{fastflightcubeSPOT}
\alias{fastflightcubeSPOT}
\title{Fast flight phase for the cube method modified}
\usage{
fastflightcubeSPOT(X, pik, order = 1, comment = TRUE)
}
\arguments{
\item{X}{matrix of auxiliary variables.}

\item{pik}{vector of inclusion probabilities.}

\item{order}{order to rearrange the data. Default 1}

\item{comment}{bool, if comment should be written.}
}
\value{

}
\description{
implementation modified from the package sampling.
}
\examples{
\dontrun{

# Matrix of balancing variables
X=cbind(c(1,1,1,1,1,1,1,1,1),c(1,2,3,4,5,6,7,8,9))
# Vector of inclusion probabilities.
# The sample size is 3.
pik=c(1/3,1/3,1/3,1/3,1/3,1/3,1/3,1/3,1/3)
# pikstar is almost a balanced sample and is ready for the landing phase
pikstar=fastflightcube(X,pik,order=1,comment=TRUE)
round(pikstar,9)





rm(list = ls())
set.seed(1)
eps <- 1e-13
library(Matrix)
N <- 300
Pik <- matrix(c(sampling::inclusionprobabilities(runif(N),70),
sampling::inclusionprobabilities(runif(N),50),
sampling::inclusionprobabilities(runif(N),30)),ncol = 3)
X <- PM(Pik)$PM
pik <- PM(Pik)$P
dim(X)
order = 2
EPS = 1e-11

system.time(test1 <- fastflightcubeSPOT(X,pik,order = 2))
system.time(test2 <- fastflightcubeSPOT(X,pik,order = 1))
system.time(test3 <- sampling::fastflightcube(X,pik, order = 2))
system.time(test4 <- BalancedSampling::flightphase(pik,X))


}
}
