% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jump.R
\name{jump}
\alias{jump}
\title{Internal function of algofastflightcube}
\usage{
jump(X, pik)
}
\arguments{
\item{X}{matrix of auxiliary variables}

\item{pik}{vector of inclusion probabilities}
}
\value{
update pik
}
\description{
Internal function of algofastflightcube
}
