% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mat_as_arma}
\alias{mat_as_arma}
\title{NumericMatrix matrix to Arma}
\usage{
mat_as_arma(x)
}
\arguments{
\item{x}{a Matrix}
}
\value{
same matrix but with arma type.
}
\description{
NumericMatrix matrix to Arma
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
