% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mat_as_Numeric}
\alias{mat_as_Numeric}
\title{Arma matrix to NumericMatrix}
\usage{
mat_as_Numeric(x)
}
\arguments{
\item{x}{a Matrix}
}
\value{
same matrix but with NumericMatrix type.
}
\description{
Arma matrix to NumericMatrix
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
