% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{onestepfastflightcube}
\alias{onestepfastflightcube}
\title{onestepfastflightcube}
\usage{
onestepfastflightcube(prob, Bm)
}
\arguments{
\item{prob}{vector of inclusion probabilities}

\item{Bm}{matrix of size (p x p+1) (transpose from the original)}
}
\value{
updated inclusion probabilities
}
\description{
one step of the fast flight cube. Direct implementation of the package BalancedSampling from Anton Grafström
}
