% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orfs.R
\encoding{UTF-8}
\name{PM}
\alias{PM}
\title{ORFS method}
\usage{
PM(Pik, tol = 1e-09)
}
\description{
ORFS method
}
\examples{
N <- 200
Pik <- matrix(c(sampling::inclusionprobabilities(runif(N),20),
sampling::inclusionprobabilities(runif(N),30),
sampling::inclusionprobabilities(runif(N),40)),ncol = 3)
X <- PM(Pik)
}
