% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rowSumsRcpp}
\alias{rowSumsRcpp}
\title{rowSums in Rcpp}
\usage{
rowSumsRcpp(x)
}
\arguments{
\item{x}{Matrix X}
}
\value{
rowSums
}
\description{
rowSums in Rcpp
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
