% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rrefArmaExport}
\alias{rrefArmaExport}
\title{reduced row echelon form arma implementation}
\usage{
rrefArmaExport(M)
}
\arguments{
\item{M}{matrix}
}
\value{
NULL (transform matrix)
}
\description{
reduced row echelon form arma implementation
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
