% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rrefBal}
\alias{rrefBal}
\title{reduced row echelon fom}
\usage{
rrefBal(M)
}
\arguments{
\item{M}{matrix}
}
\value{
NULL (transform matrix)
}
\description{
reduced row echelon fom
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
