% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplecubeSPOT.R
\name{samplecubeSPOT}
\alias{samplecubeSPOT}
\title{Title}
\usage{
samplecubeSPOT(X, pik, order = 1, comment = TRUE, method = 1)
}
\arguments{
\item{method}{}
}
\value{

}
\description{
Title
}
\examples{

N=1000
p=7
set.seed(1)
x=rnorm(N*p,10,3)
# random inclusion probabilities
pik= runif(N)
X=array(x,c(N,p))
X=cbind(cbind(X,rep(1,times=N)),pik)
pikfin = sampling::samplecube(X,pik,1,TRUE)
pikfin2 = samplecubeSPOT(X,pik,1,TRUE)






rm(list = ls())
set.seed(1)
eps <- 1e-13
library(Matrix)
N <- 200
Pik <- matrix(c(sampling::inclusionprobabilities(runif(N),70),
sampling::inclusionprobabilities(runif(N),50),
sampling::inclusionprobabilities(runif(N),30)),ncol = 3)
X <- PM(Pik)$PM
pik <- PM(Pik)$P
dim(X)
order = 2
method = 1
comment = TRUE
EPS = 1e-11

s <- samplecubeSPOT(X,pik,order = 2)




}
