% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{samplen}
\alias{samplen}
\title{All samples of fixed size}
\usage{
samplen(N, n)
}
\arguments{
\item{N}{population size}

\item{n}{sample size}
}
\value{
a matrix of all samples of fixed size.
}
\description{
description Rosetta code combn
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
