\name{st03}
\alias{st03}
\title{Data from the 2003 Sampling Theory class}
\description{
Data gathered from 15 students in the 2003 Sampling Theory class at the %%@
University of St Andrews. Rows are students; columns of data are named as %%@
follows:
 \itemize{
\item{"age"}{Student's age}
\item{"height"}{Student's height (in ft)}
\item{"weight"}{Student's weight (in stone)}
\item{"math.grade"}{Student's grade for 2nd year maths (there are some NAs)}
\item{"money.guess"}{Student's guess of the amount of money they were carrying}
\item{"money.act"}{Actual amount of money student was carrying}
\item{"exp.grade"}{Expected grade for Sampling Theory}
\item{"yr.study"}{Year of study}
 }
}
\usage{data(st03)}
\format{A data frame}
\source{2003 Sampling Theory class, University of St Andrews}
\references{None}
\keyword{datasets}