\name{st99}
\alias{st99}
\title{Data from the 1999 Sampling Theory class}
\description{
  Data gathered from 11 students in the 1999 Sampling Theory class at the University of St Andrews. Rows %%@
    are students; columns of data are named as follows:
    \itemize{
      \item{"age"}{Student's age}
                   \item{"height"}{Student's height (in ft)}
      \item{"weight"}{Student's weight (in stone)}
                      \item{"math.grade"}{Student's grade for 2nd year maths (there are some NAs)}
      \item{"money.guess"}{Student's guess of the amount of money they were carrying}
                           \item{"money.act"}{Actual amount of money student was carrying}
                           \item{"exp.grade"}{Expected grade for Sampling Theory}
                           \item{"yr.study"}{Year of study}
      }
                      }
                           \usage{data(st99)}
                           \format{A data frame}
                           \source{1999 Sampling Theory class, University of St Andrews}
                       \references{None}
\keyword{datasets}
                           