  \name{stock}
                           \alias{stock}
                           \title{Valus of six shares over a period of 760 days}
                           \description{
                           The values of six shares on Wall Street on each of 760 consecutive %%@
                           da6ys. Rows are days; columns of data are named as follows:
                           \itemize{
                           \item{"day.num"}{billionaire number (1 to 225)}
                           \item{"share1"}{Value of share number 1}
                           \item{"share2"}{Value of share number 2}
                           \item{"share3"}{Value of share number 3}
                           \item{"share4"}{Value of share number 4}
                           \item{"share5"}{Value of share number 5}
                           \item{"share6"}{Value of share number 6}
                           \item{"day"}{Day of the month}
                           \item{"month"}{Month: 1 to 12}
                           \item{"year"}{Year: 88 to 91}
                           }
                           }
                           \usage{data(stock)}
                           \format{A data frame}
                           \source{forgotten}
                           \references{None}
\keyword{datasets}


