% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{svdArma}
\alias{svdArma}
\title{svd in arma}
\usage{
svdArma(x)
}
\arguments{
\item{x}{Matrix X}
}
\value{
list of element of svd
}
\description{
svd in arma
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
