% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemporalCube.R
\name{TemporalCube}
\alias{TemporalCube}
\title{Cube method for temporal samples selection}
\usage{
TemporalCube(S, P, R, tol = 1e-08)
}
\arguments{
\item{S}{a matrix that contains variables on which the sample must be balanced. See Details.}

\item{P}{a vector of probabilities of select samples.}

\item{R}{a vector that specify to which unit each sample belongs.}

\item{tol}{the tolerance parameter. Default value is 1e-8.}
}
\value{
\code{PP} the balanced sample selected (a vector of 0s and 1s with the same size as \code{P}).
}
\description{
Select one temporal sample for each unit using the Cube method.
The function \code{\link[sampling:samplecube]{samplecube}} from the package \code{sampling} is used.
}
\details{
Balancing constraints considered in \code{S} allow first to select exactly one temporal sample for each unit considered
and then to verify constraint of fixed samples size.
}
\examples{
\dontrun{

## Temporal samples of three units ##
S <- matrix(c(0,1,1,
              1,1,0,
              1,0,1,
              0,0,1,
              0,1,0,
              1,0,1,), ncol=3, byrow = T)
R <- c(1,1,2,2,3,3)
P <- c(0.2,0.8,0.6,0.4,0.5,0.5)

## Find balanced sample ##
res  <- LandTemporalPivot(S, P, R, tol = 1e-6)
res
}

}
\author{
Esther Eustache \email{esther.eustache@unine.ch}
}
