% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ukernArma}
\alias{ukernArma}
\title{kernel vector arma}
\usage{
ukernArma(Bm)
}
\arguments{
\item{Bm}{matrix}
}
\value{
a null vector
}
\description{
kernel vector arma
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
