\name{ukrain}
\alias{ukrain}
\title{Annual rainfall in the UK}
\description{
  Annual rainfall at each of 43 weather stations. Rows are stations; %%@
    columns of data are named as follows:
    \itemize{
      \item{"station"}{Weather station number: 1 to 43}
      \item{"rain"}{Annual rainfall in inches}
      \item{"H"}{Mean measurement of H+ in rain}
      \item{"SO4"}{Mean measurement of SO4+ in rain}
      \item{"NO3"}{Mean measurement of NO3- in rain}
      \item{"NH4"}{Mean measurement of NH4+ in rain}
      \item{"lon"}{Relative longitude coordinate (reference point is in SE %%@
                                                    of UK)}
      \item{"latn"}{Relative latitude coordinate (reference point is in SE %%@
                                                    of UK)}
      \item{"dist.ects"}{Distance of station from the east coast of the UK %%@
                           at the same latitude as the station}
    }
}
\usage{data(ukrain)}
\format{A data frame}
\source{forgotten}
\references{None}
\keyword{datasets}
