\name{worms}
\alias{worms}
\title{Numbers of worms in a field.}
\description{
  Numbers of worms in each square of a 20m by 20m grid spanning a field. %%@
    Rows are squares; columns of data are named as follows:
    \itemize{
      \item{"square"}{Square number: 1 to 400}
      \item{"count"}{Number of worms in the square}
      \item{"north"}{Distance north at the northernmost edge of the square, %%@
                       in metres}
      \item{"west"}{Distance est at the westernmost edge of the square, in %%@
                      metres}
      \item{"alt"}{Altitude of the centre of the square, in metres}
    }
}
\usage{data(worms)}
\format{A data frame}
\source{pers. commn}
\references{None.}
\keyword{datasets}
