% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orfs.R
\encoding{UTF-8}
\name{Orfs}
\alias{Orfs}
\title{ORFS method}
\usage{
Orfs(Pik, tol = 1e-09)
}
\arguments{
\item{Pik}{a matrix of temporal inclusion probabilities.
Columns of \code{Pik} correspond to the waves, and rows correspond to the units.
Inclusion probabilities can be totally unequal.}

\item{tol}{the tolerance parameter. Default value is 1e-9.}
}
\value{
\code{S} a matrix that contains temporal samples.
It is the update of \code{Pik} and contains only 0s and 1s that indicates if a unit is selected or not at each wave.
}
\description{
Select temporal samples of fixed size at each wave.
It allows to generate an optimal temporal rotation in the selection of units using the systematic sampling method.
}
\examples{
## Temporal inclusion probabilities with 3 waves and 4 units ##
Pik <- matrix(c(0.6,0.3,0.3,
                0.2,0.4,0.9,
                0.3,0.2,0.5,
                0.9,0.1,0.3), ncol = 3, byrow = TRUE)

N <- 200
Pik <- matrix(c(sampling::inclusionprobabilities(runif(N),20),
sampling::inclusionprobabilities(runif(N),30),
sampling::inclusionprobabilities(runif(N),40)),ncol = 3)

## ORFS method ##
res <- Orfs(Pik, tol = 1e-6)
res

}
\author{
Esther Eustache \email{esther.eustache@unine.ch}
}
