% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orfs.R
\name{PM}
\alias{PM}
\title{just for test some function}
\usage{
PM(Pik, tol = 1e-09)
}
\arguments{
\item{Pik}{vector of inclusion probabilities}

\item{tol}{tolerance}
}
\description{
just for test some function
}
\examples{
\dontrun{
N <- 200
Pik <- matrix(c(sampling::inclusionprobabilities(runif(N),20),
sampling::inclusionprobabilities(runif(N),30),
sampling::inclusionprobabilities(runif(N),40)),ncol = 3)
X <- PM(Pik)$PM
}
}
