% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ukern}
\alias{ukern}
\title{kernel vector}
\usage{
ukern(Bm)
}
\arguments{
\item{Bm}{matrix}
}
\value{
a null vector
}
\description{
kernel vector
}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
